chore: default time range
